+++ server/routes/index.js
@@ -20,9 +20,9 @@ router.get('/', (req, res) => {
 	res.sendFile(path.resolve('public', 'index.html'));
 });
 
-router.get('/login', (req, res) => {
-	res.sendFile(path.resolve('public','login.html'));
-});
+// router.get('/login', (req, res) => {
+// 	res.sendFile(path.resolve('public','login.html'));
+// });
 
 // Use JWT to authenticate
 router.post('/login', (req, res) => {
@@ -66,6 +66,8 @@ router.post('/login', (req, res) => {
 
 router.post('/signup', (req, res) => {
 	const {username, fullname, password, password2, email} = req.body;
+
+	/** Add unique username and unique email **/
 
 	// Validation from expressValidator
 	req.checkBody('fullname', 'Name is required').notEmpty();
